+++ lib/src/file_cache_image.dart
@@ -1,11 +1,27 @@
 import 'dart:async';
-import 'dart:typed_data';
 
 import 'package:flutter/painting.dart';
 import 'package:flutter/foundation.dart';
 import 'dart:ui' as ui show instantiateImageCodec, Codec;
+import 'package:path_provider/path_provider.dart';
+import 'package:universal_io/io.dart';
+import 'package:path/path.dart' as p;
 
 import 'file_cache.dart';
+
+class FileCacheFlutter {
+  static Future<FileCache> fromDefault({
+    Loader loader = defaultLoader,
+    String? path,
+    bool scan = false,
+  }) async {
+    if (path == null) {
+      Directory dir = await getTemporaryDirectory();
+      path = p.join(dir.path, 'cache2');
+    }
+    return FileCache.from(loader:loader,path: path, scan: scan);
+  }
+}
 
 class FileCacheImage extends ImageProvider<FileCacheImage> {
   /// Creates an ImageProvider which loads an image from the [url], using the [scale].
@@ -35,7 +51,7 @@ class FileCacheImage extends ImageProvider<FileCacheImage> {
 
   Future<ui.Codec> _loadAsync(FileCacheImage key) async {
     assert(key == this);
-    FileCache fileCache = await FileCache.fromDefault();
+    FileCache fileCache = await FileCacheFlutter.fromDefault();
 
     final Uint8List bytes = await fileCache.getBytes(key.url);
     return await ui.instantiateImageCodec(bytes);